test merge (#1)
* fix(deps): update dependency p-queue to ^6.6.1 (#1838)

Co-authored-by: Renovate Bot <bot@renovateapp.com>

* fix(deps): update dependency @babel/parser to ^7.11.4 (#1844)

Co-authored-by: Renovate Bot <bot@renovateapp.com>

* chore(deps): update dependency mock-fs to ^4.13.0 (#1845)

Co-authored-by: Renovate Bot <bot@renovateapp.com>

* chore(deps): update dependency eslint to v7.12.1 (#1851)

Co-authored-by: Renovate Bot <bot@renovateapp.com>

* chore(deps): update dependencies

* chore(deps): update dependency eslint to v7.14.0 (#1882)

Co-authored-by: Renovate Bot <bot@renovateapp.com>

* fix: prevent circular refs when params have the same type expression

Catharsis caches parse results by default; if you parse the same type expression twice, with the same options, you get the same object each time.

When the user passes the `--debug` flag, we expose the parsed type for each parameter as an enumerable property of the doclet. If two parameters used the same type expression, the resulting doclet could contain a circular reference.

This change disables the Catharsis cache, so that Catharsis returns a new object for each parsed type, which prevents circular references. As a result, this change fixes an issue with the `-X` flag, and with some JSDoc templates.

* fix: correctly track interface members

When an interface is a) defined as an ES2015 class and b) assigned to a variable, JSDoc sometimes used the wrong `longname` and `memberof` for members of the interface (specifically, for instance properties).

The root cause was that we weren't resolving `this` correctly within this type of interface. As a result, if you added a JSDoc comment to something like `this.foo = 'bar'`, the doclet for `this.foo` had the wrong `longname` and `memberof`.

Fixing that issue uncovered another issue: When we merged the constructor's doclet with the interface's doclet, we preferred the constructor's doclet. However, the constructor's doclet used the wrong `kind` in this case; we already had code to fix up the `longname` and `memberof` of the combined doclet, but not the `kind`. The fix was to prefer the interface's doclet for all properties.

Manual cherry-pick of 95e3192525310b9f1567e034c22489da3a5847a1.

* chore: merge changelog for 3.6.4-3.6.6

* chore(deps): update dependencies

* chore(deps): update dependency eslint to v7.16.0 (#1885)

Co-authored-by: Renovate Bot <bot@renovateapp.com>

* chore: update ESLint rules

* chore: migrate from Travis CI to GitHub Actions

* chore: use GitHub Actions build-status badge

* chore: disable Windows CI builds for now

* chore: fix CI badge

* fix(deps): update dependency @babel/parser to ^7.12.11 (#1888)

Co-authored-by: Renovate Bot <bot@renovateapp.com>

* chore(deps): update dependency ajv to v7 (#1889)

Co-authored-by: Renovate Bot <bot@renovateapp.com>

* chore(deps): update dependency ow to ^0.21.0

* chore(jsdoc): remove jsdoc/env from astbuilder

* chore(jsdoc): remove jsdoc/env from parser

* chore(jsdoc): remove jsdoc/env from parser tests

* chore(jsdoc): remove jsdoc/env from astnode

* refactor(jsdoc-core): add Syntax to @jsdoc/core

* Publish

 - @jsdoc/cli@0.2.4
 - @jsdoc/core@0.3.3
 - @jsdoc/eslint-config@1.1.5
 - @jsdoc/task-runner@0.1.9
 - @jsdoc/test-matchers@0.1.5
 - @jsdoc/util@0.2.3
 - jsdoc@4.0.0-dev.14

* chore: delete unnecessary .gitignore files

* chore(jsdoc-eslint-config): add missing .npmignore file

* refactor(jsdoc-core): move syntax to new @jsdoc/parse package

* Publish

 - @jsdoc/cli@0.2.5
 - @jsdoc/core@0.4.0
 - @jsdoc/eslint-config@1.1.6
 - @jsdoc/parse@0.1.0
 - @jsdoc/task-runner@0.1.10
 - @jsdoc/test-matchers@0.1.6
 - @jsdoc/util@0.2.4
 - jsdoc@4.0.0-dev.15

* refactor(jsdoc): use Syntax from @jsdoc/parse

* refactor: move ast-builder and ast-node to @jsdoc/parse

BREAKING CHANGE: Modules no longer exist in jsdoc package.

* refactor(jsdoc): move tag and inline to new @jsdoc/tag package

* Publish

 - @jsdoc/parse@0.1.1
 - @jsdoc/tag@0.1.0
 - jsdoc@4.0.0-dev.16

* chore(jsdoc-tag): add missing .npmignore contents

* chore(deps): update actions/setup-node action to v2

* fix(deps): update dependency emittery to ^0.8.0

* chore(deps): update dependency eslint to v7.17.0

* chore(deps): update dependency ajv to ^7.0.3

* fix(deps): update dependency emittery to ^0.8.1

* fix(deps): update dependency ow to ^0.22.0

* fix(deps): update dependency dependency-graph to ^0.10.0

* chore(deps): update dependency eslint to v7.18.0

* chore(deps): update dependency jasmine to ^3.6.4

* fix(deps): update dependency ow to ^0.23.0

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
Co-authored-by: Renovate Bot <bot@renovateapp.com>
Co-authored-by: Jeff Williams <jeffrey.l.williams@gmail.com>